python-kafka-example/requirements.txt
python-dotenv==0.20.0
confluent_kafka
